Extract directory enumeration partial
